Update LineItem to use Long IDs

+++ app/src/main/java/com/example/struku/domain/model/LineItem.kt
@@ -4,8 +4,8 @@ package com.example.struku.domain.model
  * Model domain untuk item baris dalam struk
  */
 data class LineItem(
-    val id: Int = 0,
-    val receiptId: Int = 0,
+    val id: Long = 0,
+    val receiptId: Long = 0,
     val description: String,
     val quantity: Int = 1,
     val unitPrice: Double? = null,
